Unified logging and console output
This removes the need to have duplicate logging and print calls by
initializing a console log handler along with the optional file log
handler. Many logging and print calls are consolidated and the levels
set to more useful values.

If stderr is a TTY the coloredlogs module will be used to make the
output easier to read.

The new --verbose argument allows increasing the log level for all
handlers.

The new --log-file argument allows setting the filename for the
now-optional log file.